Add first working version of the carousel
Had to remove the slide class from it. The animation is broken for some
reason, I suspect the tutorial's removal of the indicators but am too
tired to keep iterating. For now, the rougher animation will have to do.

All we really need to do next time we advance the project is add the
content for the 2 extra clients