Fix carousel. Add filler text to testimonials. Minor style changes
Previous implementation of the carousel barely worked and only had 2
testimonials. It now has been fixed by lifting code from the proper
bootstrap docs page (previous iteration base code came from a different
source) and tweaked from there to match the previous style we had. The
slides have no gap during the transition animation but that's a
limitation of the way bootstrap handles carousels. Would need research
and work in the inner settings of bootstrap to edit.

The images are already working, but I still need to remember to add
client-2 and client-3 with proper images of other football players to
keep up with the easter egg theme of the first pic.