Swap accordion icons in faq section
Had to experiment a bit with the imported code from bootstrap icons page
to make it work. Turns out the stroke settings in the !default code mess
up with the icons, making them larger and thus clip through the bounding
box. The fix was to delete the stroke lines and instead have the colors
we wanted come through with the fill setting.